+++ test/providers/pythonCompletion.test.ts
@@ -1,14 +1,12 @@
 import fs from 'fs';
 import path from 'path';
-import { Logger } from 'winston';
+import type { Logger } from 'winston';
 import { getCache, isCacheEnabled } from '../../src/cache';
-import { importModule } from '../../src/esm';
 import logger from '../../src/logger';
 import { PythonProvider } from '../../src/providers/pythonCompletion';
 import { runPython } from '../../src/python/pythonUtils';
 import { parsePathOrGlob } from '../../src/util';
-import { isJavascriptFile } from '../../src/util/file';
-import { loadFileReference, processConfigFileReferences } from '../../src/util/fileReference';
+import { processConfigFileReferences } from '../../src/util/fileReference';
 
 // Mock dependencies
 jest.mock('fs');
@@ -43,7 +41,7 @@ describe('PythonProvider', () => {
     jest.resetAllMocks();
 
     // Mock logger methods with jest.fn()
-    (logger.debug as jest.Mock).mockImplementation(
+    (jest.mocked(logger.debug)).mockImplementation(
       () =>
         ({
           debug: jest.fn(),
@@ -53,7 +51,7 @@ describe('PythonProvider', () => {
         }) as unknown as Logger,
     );
 
-    (logger.error as jest.Mock).mockImplementation(
+    (jest.mocked(logger.error)).mockImplementation(
       () =>
         ({
           debug: jest.fn(),
@@ -193,8 +191,7 @@ describe('PythonProvider', () => {
       provider['configReferencesProcessed'] = false;
 
       // Mock the entire executePythonScript method
-      provider['executePythonScript'] = jest
-        .fn()
+      jest.spyOn(provider, 'executePythonScript').mockImplementation()
         .mockResolvedValue({ output: 'API result', cached: false });
 
       jest.mocked(runPython).mockResolvedValue({ output: 'API result' });
